+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.6.0]
+## [0.6.0] - 2020-06-09
 
 - Adds `cookie` middleware.
 - Adds `accept` functions to do content negotiation.